refactor(warm): extract the source agnostic parts of the warm pipeline
Job registries and log helpers (job-control), target VM shell and block volume provision
with the thick pre zero (target-provision), extent application with progress (apply), the
confirmed power off wait (power-off) and the attach/boot/verify tail (finish) move to
their own modules so a second warm pipeline can share them. runWarmMigration keeps its
flow and log lines; every previously exported name is re-exported.
